Uploaded the new League of Legends Summoner API
This is a new project for a personal discord community. This is a WIP, but aims to lookup player data and return it in a simple format. This file only works in the CLI for now, I have a discord bot in the works that will be more fleshed out. This is likely going to be the last change for this file (CLI only).

This script contains:
- Player data (Level, ranked stats)
- Match history (Includes all players from each game, bubt has a section for the user's stats)
- Append tags to players based on certain performances
- Built-in smurf detector (very basic but will be slowly improved)